style: HandleReponse 코드로 변경 [IOS-35]
